Rename storage2 mod to storage
