+++ src/storage/memcache.rs
